Plantilla WEB y agregador a winform

+++ ClassLibrary1/Compañia.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Compañia
     {
-        public short ID { get; set; }
+        public int ID { get; set; }
         public string Nombre { get; set; }
         public override string ToString()
         {
